Potential fix
Capcom changed the game so the variable for dying triggers true at the end of Mia's tape. This was causing the videotape end split to not trigger. As a result, the potential fix is to remove the check to make sure the player isn't dead. As a result, however, the Videotape End split will fire if you die during Mia's tape. It's a forced "bug".